Rename {is_error_code_enum => error_code_enum}.hpp
