+++ libcaf_core/caf/error.hpp
@@ -7,8 +7,8 @@
 #include "caf/detail/comparable.hpp"
 #include "caf/detail/core_export.hpp"
 #include "caf/error_code.hpp"
+#include "caf/error_code_enum.hpp"
 #include "caf/fwd.hpp"
-#include "caf/is_error_code_enum.hpp"
 #include "caf/message.hpp"
 #include "caf/none.hpp"
 #include "caf/type_id.hpp"
