+++ libcaf_core/caf/error_code_enum.hpp
